requirements.txt
python-binance
flask
sqlalchemy
ccxt